Add ContentViewResolver service for resolving ContentView instances
